Extract shortcuts into multiple files
