+++ safety_tests/tests.py
@@ -5,7 +5,7 @@ from django.db import models
 from django.test import TransactionTestCase
 from django_fake_model import models as f
 
-from safety.shortcuts import set_perm, has_perm, lift_perm, create_object_group, delete_object_group, \
+from safety.utils import set_perm, has_perm, lift_perm, create_object_group, delete_object_group, \
     add_user_to_object_group, remove_user_from_object_group, get_users_with_perms, get_groups_with_perms, \
     retrieve_object_group, get_objects_for_entity
 from safety_tests.models import FakePost
